refactor: move WritingAssistant types to commons
Move WritingAssistant interface, WritingTone enum, WritingAssistantStatus
sealed class, and WritingResult data class from amethyst.service.ai to
commons.service.ai for KMP shared access.
Update all consumers (play, fdroid, main) to import from commons directly.
